requirements.txt
redis
rq
google-cloud-happybase